partially use `skipFinalSub` in EC sum
We cannot use it in all the same places as for the CPU
implementation. There is some kind of hidden bug here, related to
remaining state.

See the notes in the code, but essentially if we run the
`tests/gpu/t_ec_sum.nim` test case after a certain number of
iterations of

let a, b be EC points
res = a
while true:
  res = ec sum(res, b)

the code will fail to match the expected CPU result for the same
operation.

Maybe related to the additional bits available for the big ints
storing some problematic data?