Nvidia MSM proof of concept (serial) (#480)
* wrap `execCudaImpl` macro logic in a block

Otherwise we run into problems if we have two execs in the same scope.

* add more EC Jac operations to helper templates

* do not quit on failure in NvidiaAssembler destructor

A failure in the check from the destructor almost certainly means that
we destroyed early, due to an exception. We don't want to hide the
exception, hence we don't quit.

* add CurveDescriptor fields for LLVM type for Fr, scalars for MSM

* [LLVM] add `isPointerTy` helper to determine if type is a pointer

* [tests] add sanity test for adding neutral EC element to EC sum

* store EC order bit width in CurveDescriptor

* make `store` for `ValueRef` safer by checking for pointer-ness

Also adds `storePtr` if user really wants to store a pointer

* forbid `=copy` on Array, likely *not* what user wants

Easy to introduce bugs by thinking one stores, when in fact one just
copies the reference.

* allow access read/write of `Array` using `ValueRef`

* add `FieldScalar`, `FieldScalarArray`, `EcAffArray`, `EcAffArray`

- for safer handling of multiple EC points in different coordinates
- separate logic of elements of Fp (`Field`) from those of Fr (`FieldScalar`)

* extend doc string of `compile` taking a string

* add ConstantValue, MutableValue wrappers around ValueRef

Dealing with ValueRef and the fact that pointers are now opaque in
LLVM is extremely annoying. So here are 2 types that wrap the LLVM
values with their respective underlying types which also provide
easier load / write access.

* add `llvmFor` macro that produces code for a for loop in LLVM

* add helpers for arithmetic, boolean logic for ValueRef, M/CValue

* add `llvmIf` to generate code for if statements

It _wraps around_ a full if statement.

* add `to` type conversion helper which extends/truncates int types

* use `llvmForCountdown` in `genFpNsqrRt` instead of fixed countdown logic

* add `getWindowAt` helper required for baseline MSM implementation

* add serial MSM implementation for Nvidia using bucket method

This implementation is a bit of a proof of concept and playground to
investigate how easily we can generate code on the LLVM target with
the help of Nim macros.

* [tests] add mini test case for MSM on Nvidia

* whoops, revert local change to test CT error on `=copy`