[cuda] explicitly support constants, mapped to `__constant__`
One can either define a Nim `const` for a variable that is already a
const at the Nim compile time or use
```nim
var foo {.constant.}: theType
```
if one wishes to copy to the symbol before kernel execution. This is
useful for global constants that are not filled in at CUDA compile
time, but before execution. For example:

```nim
  # Filled with `copyToSymbol` at runtime from host!
  var rc16 {.constant.}: array[30, array[BABYBEAR_WIDTH, BigInt]]
  var matInternalDiagM1 {.constant.}: array[BABYBEAR_WIDTH, BigInt]
  var montyInverse {.constant.}: BigInt
```

And in the host code:
```nim
  var nvrtc = initNvrtc(CudaCode)
  nvrtc.compile()
  nvrtc.getPtx()
  nvrtc.load()

  var p2bb = Poseidon2BabyBear.init()

  # copy Poseidon2 constants to CUDA kernel
  nvrtc.copyToSymbol("rc16", p2bb.rc16)
  nvrtc.copyToSymbol("matInternalDiagM1", p2bb.matInternalDiagM1)
  nvrtc.copyToSymbol("montyInverse", p2bb.montyInverse)
```