revert default of `passStructByPointer` in `execCudaImpl`
This argument decides whether an object type is passed via the host
pointer *or* copied to a `CUdeviceptr`.

We disabled it in the past, because it caused issues for when
launching kernels with multiple threads.

However, the design as of right now breaks for more basic cases if
this is set to `false`.

The real fix will rely on comparing the data passed into `execute`
with the signature of the kernel we pass in, which will be done in the
future.

For the moment you simply need to be careful to call the kernel
correctly. If you want things to be passed by copy (i.e. copied
manually to a device pointer and with a kernel of an input as a
pointer), you should define the kernel to take a pointer and then
simply pass in a pointer to `execute` (or pass in the non pointer data
into the `res` argument)