implement lifting of struct pointer fields
This gives us the convenience of passing around a struct with a
pointer field, while preserving the restrictions of WebGPU. Because of
the fact that storage buffers are global anyway, we can "pretend" the
pointers are part of an object and simply replace them by the global
when the fields are used.

We throw a CT error if one tries to assign a non storage buffer
pointer to a field.