allow "pulling in" procs from outside `cuda` scope
The generic logic is essentially what we need to pull in a proc
defined outside the scope of the `cuda` macro into the
code. Essentially when we encounter a function that is not known to
us yet, we simply look up its implementation from the symbol.

This is the first step towards untying the current `cuda` macro code
from relying on having _everything_ be defined under that macro.

In the future the idea is that essentially only the `{.global.}` procs
strictly need to be defined in the macro. These then pull in
everything they need (i.e. everything that is used and has been
checked by the compiler to be used).