improve type deduction for types that look generic
I.e. an array type with a constant in place for the array
size (i.e. *not* a `static int` generic argument) will look like a
generic in the sense that depending on how you look at the type with
`getType*`, you will end up seeing `Ident "Foo"` for the constant.

Therefore, we allow array length determination to fail and return
`gtInvalid`. If we see that happens for a variable or parameter, we
try to determine the type from the symbol of the parameter or
variable. For some reason that tends to have the fully instantiated
type.

However, for *return types* this is not possible, as there is no easy
symbol to look at (unless one were to go into the proc body and look
at symbol that is being returned). Therefore, if we encounter this
type of array as a return type (NOTE: most backends won't anyway allow
to return arrays, but that's beside the point), we treat the function
as a generic and look up its type when we encounter it in a `gpuCall`.
Then, the type will be fully instantiated again.