[cuda] for now emit `constexpr` for a `gpuConstexpr` (i.e. Nim `const`)
See the added TODO. The main point is that if we write things like

```nim
  const M = toBigInt(bigintToUint32Limbs(T.getModulus))
```

we want to have the Nim compiler evaluate the RHS at compile time. If
we were to make the user write

```
  var M {.constant.} = toBigInt(bigintToUint32Limbs(T.getModulus))
```

instead the Nim compiler would evaluate the RHS at runtime.

We _could_ force the user to write

```nim
  const M {.constant.} = toBigInt(bigintToUint32Limbs(T.getModulus))
```

instead though.

It is no problem however, to emit `__constant__` if it's a global and
`constexpr` for a local (where `__constant__` is anyhow forbidden in
CUDA). The only minor annoyance is that _maybe_ someone wants to emit
`constexpr` also in global scope.

_Maybe_ we'll go with a design that does the above by default but
allows you to overwrite it via

```nim
  const M {.constant.} = toBigInt(bigintToUint32Limbs(T.getModulus))
```
-> explicitly force `__constant__`

```nim
  const M {.constexpr.} = toBigInt(bigintToUint32Limbs(T.getModulus))
```
-> explicity force `constexpr` instead.