implement initial WebGPU / WGSL backend, split GPU compiler up (#564)
* implement initial WebGPU / WGSL backend, split GPU compiler up

Implements an initial implementation of a WebGPU backend (WGSL).

Helper functions to execute code are not part of this, as they
require HTML and JavaScript code.

It includes a refactor of the existing NVRTC compiler code into
smaller pieces

* change mutation logic of type replacements for WGSL

As we nowadays have support for UncheckedArray as a specific type, we
don't need any special rules anymore. Essentially for `ptr T`
arguments we just emit `T`. If one wants to emit `array<T>`, just use
a `ptr UncheckedArray[T]` and it'll work.

* remove TODO comment

The answer to the question raised in the comment is two fold.

Take a snippet like this:
```nim
let Example = toGpuAst:
  const NUM_WORKGROUPS = 1
  proc other(x: ptr uint32, y: uint32) {.device.} =
    x[] = y

  proc main(x: ptr uint32) {.global.} =
    other(x, 5'u32)
```

1. We only call `injectAddressOf` for *globals* anyway!
2. Even if we were not, the check `n.dOf.iSym in ctx.globals` would
   *not* trigger for the symbol `x` in the `other` proc above. They
   are two distinct symbols!

* change `gpuProc` length to be 1

We only yield `pBody`, so that means the length is one.

* always map `int` to `gtInt32` independent of backend

* rename `nvrtc_field_ops` to `gpu_field_ops`