Support more uint literals, simplify registering types for GPU, regular finite field `pow` typo fix (#583)
* fix typo in in-place `pow` for finite fields

* support uint literals 'u16, 'u32, 'u64

* simplify collecting types by doing it in `nimToGpuType`

Instead of having calls to `maybeAddType` all over the place, we now
pass the GPU context into the `nimToGpuType` function and call the
helper to register types in there. That makes the most sense and is
the cleanest.

Initially I was against this, because it was nice that ~nimToGpuType~
was a pure function without dependence on any kind of context
state. But the upside is worth it and we don't really need the
function in another context anyway.